main/main/django/viewsets.py
from main.django.serializers import AlbumSerializer
from main.django.models import Album
from rest_framework import viewsets


class AlbumViewSet(viewsets.ModelViewSet):
    queryset = Album.objects.all()
    serializer_class = AlbumSerializer
