main/requirements.txt
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
Automat==22.10.0
binaryornot==0.4.4
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cookiecutter==2.6.0
cryptography==42.0.8
cssselect==1.2.0
defusedxml==0.7.1
Django==5.0.7
django-browser-reload==1.13.0
django-filter==24.3
django-tailwind==3.8.0
djangorestframework==3.15.2
filelock==3.15.4
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
parsel==1.9.1
plotly==5.23.0
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
Scrapy==2.11.2
scrapy-djangoitem==1.1.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tenacity==9.0.0
text-unidecode==1.3
tldextract==5.1.2
Twisted==24.3.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
w3lib==2.2.1
zope.interface==6.4.post2
